possible fix for sagemaker issue

+++ hub/core/storage/s3.py
@@ -7,7 +7,6 @@ from botocore.session import ComponentLocator
 from hub.client.client import HubBackendClient
 from hub.core.storage.provider import StorageProvider
 from hub.util.exceptions import S3DeletionError, S3GetError, S3ListError, S3SetError
-import hub
 
 
 class S3Provider(StorageProvider):
@@ -71,15 +70,32 @@ class S3Provider(StorageProvider):
         """
         self.check_readonly()
         self._check_update_creds()
+        path = posixpath.join(self.path, path)
+        content = bytearray(memoryview(content))
         try:
-            path = posixpath.join(self.path, path)
-            content = bytearray(memoryview(content))
             self.client.put_object(
                 Bucket=self.bucket,
                 Body=content,
                 Key=path,
                 ContentType="application/octet-stream",  # signifies binary data
             )
+        # catch expired token error
+        except botocore.exceptions.ClientError as err:
+            if err.response["Error"]["Code"] == "ExpiredToken" and self.loaded_creds_from_environment:
+                self._locate_and_load_creds()
+                self._set_s3_client_and_resource()
+                try:
+                    self.client.put_object(
+                        Bucket=self.bucket,
+                        Body=content,
+                        Key=path,
+                        ContentType="application/octet-stream",  # signifies binary data
+                    )
+                except Exception as err:
+                    raise S3SetError(err)
+            else:
+                raise S3SetError(err)
+                    
         except Exception as err:
             raise S3SetError(err)
 
@@ -98,17 +114,29 @@ class S3Provider(StorageProvider):
             ReadOnlyError: If the provider is in read-only mode.
         """
         self._check_update_creds()
+        path = posixpath.join(self.path, path)
         try:
-            path = posixpath.join(self.path, path)
             resp = self.client.get_object(
                 Bucket=self.bucket,
                 Key=path,
             )
-            return resp["Body"].read()
+            return resp["Body"].read()            
         except botocore.exceptions.ClientError as err:
             if err.response["Error"]["Code"] == "NoSuchKey":
                 raise KeyError(err)
-            raise S3GetError(err)
+            elif err.response["Error"]["Code"] == "ExpiredToken" and self.loaded_creds_from_environment:
+                self._locate_and_load_creds()
+                self._set_s3_client_and_resource()
+                try:
+                    resp = self.client.get_object(
+                        Bucket=self.bucket,
+                        Key=path,
+                    )
+                    return resp["Body"].read()
+                except Exception as err:
+                    raise S3GetError(err)
+            else:
+                raise S3GetError(err)
         except Exception as err:
             raise S3GetError(err)
 
@@ -125,9 +153,20 @@ class S3Provider(StorageProvider):
         """
         self.check_readonly()
         self._check_update_creds()
+        path = posixpath.join(self.path, path)
         try:
-            path = posixpath.join(self.path, path)
             self.client.delete_object(Bucket=self.bucket, Key=path)
+        # catch expired token error
+        except botocore.exceptions.ClientError as err:
+            if err.response["Error"]["Code"] == "ExpiredToken" and self.loaded_creds_from_environment:
+                self._locate_and_load_creds()
+                self._set_s3_client_and_resource()
+                try:
+                    self.client.delete_object(Bucket=self.bucket, Key=path)
+                except Exception as err:
+                    raise S3DeletionError(err)
+            else:
+                raise S3DeletionError(err)
         except Exception as err:
             raise S3DeletionError(err)
 
@@ -144,16 +183,28 @@ class S3Provider(StorageProvider):
         try:
             # TODO boto3 list_objects only returns first 1000 objects
             items = self.client.list_objects_v2(Bucket=self.bucket, Prefix=self.path)
-            if items["KeyCount"] <= 0:
-                return set()
-            items = items["Contents"]
-            names = [item["Key"] for item in items]
-            # removing the prefix from the names
-            len_path = len(self.path.split("/")) - 1
-            names = {"/".join(name.split("/")[len_path:]) for name in names}
-            return names
+        # catch expired token error
+        except botocore.exceptions.ClientError as err:
+            if err.response["Error"]["Code"] == "ExpiredToken" and self.loaded_creds_from_environment:
+                self._locate_and_load_creds()
+                self._set_s3_client_and_resource()
+                try:
+                    items = self.client.list_objects_v2(Bucket=self.bucket, Prefix=self.path)
+                except Exception as err:
+                    raise S3ListError(err)
+            else:
+                raise S3ListError(err)
         except Exception as err:
             raise S3ListError(err)
+        
+        if items["KeyCount"] <= 0:
+            return set()
+        items = items["Contents"]
+        names = [item["Key"] for item in items]
+        # removing the prefix from the names
+        len_path = len(self.path.split("/")) - 1
+        names = {"/".join(name.split("/")[len_path:]) for name in names}
+        return names
 
     def __len__(self):
         """Returns the number of files present at the root of the S3Provider. This is an expensive operation.
@@ -181,8 +232,22 @@ class S3Provider(StorageProvider):
         self.check_readonly()
         self._check_update_creds()
         if self.resource is not None:
-            bucket = self.resource.Bucket(self.bucket)
-            bucket.objects.filter(Prefix=self.path).delete()
+            try:
+                bucket = self.resource.Bucket(self.bucket)
+                bucket.objects.filter(Prefix=self.path).delete()
+            # catch expired token error
+            except Exception as e:
+                if self.loaded_creds_from_environment:
+                    self._locate_and_load_creds()
+                    self._set_s3_client_and_resource()
+                    try:
+                        bucket = self.resource.Bucket(self.bucket)
+                        bucket.objects.filter(Prefix=self.path).delete()
+                    except Exception as err:
+                        raise S3DeletionError(err)
+                else:
+                    raise S3DeletionError(e)
+
         else:
             super().clear()
 
@@ -198,6 +263,7 @@ class S3Provider(StorageProvider):
             self.expiration,
             self.tag,
             self.token,
+            self.loaded_creds_from_environment
         )
 
     def __setstate__(self, state):
@@ -211,6 +277,7 @@ class S3Provider(StorageProvider):
         self.expiration = state[7]
         self.tag = state[8]
         self.token = state[9]
+        self.loaded_creds_from_environment = state[10]
 
         self._initialize_s3_parameters()
 
@@ -239,9 +306,11 @@ class S3Provider(StorageProvider):
         self.client_config = botocore.config.Config(
             max_pool_connections=self.max_pool_connections,
         )
+        self.loaded_creds_from_environment = False
 
         if self.aws_access_key_id is None and self.aws_secret_access_key is None:
             self._locate_and_load_creds()
+            self.loaded_creds_from_environment = True
 
         self._set_s3_client_and_resource()
